+++ tests/net/net_pkt/prj.conf
@@ -1,21 +1,14 @@
-CONFIG_NETWORKING=y
 CONFIG_NET_TEST=y
-CONFIG_NET_IPV6=y
-CONFIG_NET_BUF=y
-CONFIG_NET_BUF_LOG=y
-CONFIG_MAIN_STACK_SIZE=2048
-CONFIG_NET_PKT_RX_COUNT=4
-CONFIG_NET_PKT_TX_COUNT=4
-CONFIG_NET_BUF_RX_COUNT=34
-CONFIG_NET_BUF_TX_COUNT=34
-# The data size is calculated to be this, do not change
-# it without fixing the tests.
-CONFIG_NET_BUF_DATA_SIZE=100
-CONFIG_NET_LOG=y
-CONFIG_ENTROPY_GENERATOR=y
-CONFIG_TEST_RANDOM_GENERATOR=y
 CONFIG_ZTEST=y
 
-# Do not print any errors, they just confuse the sanitychecker
-# even if there is no problem with the test.
-CONFIG_LOG_MAX_LEVEL=0
+CONFIG_NETWORKING=y
+CONFIG_NET_L2_ETHERNET=y
+CONFIG_NET_IPV4=y
+CONFIG_NET_IPV6=n
+CONFIG_NET_ARP=y
+CONFIG_NET_UDP=y
+CONFIG_NET_BUF_LOG=y
+CONFIG_NET_LOG=y
+
+CONFIG_ENTROPY_GENERATOR=y
+CONFIG_TEST_RANDOM_GENERATOR=y
