--- tests/net/net_pkt_new/prj.conf
@@ -1,14 +1,0 @@
-CONFIG_NET_TEST=y
-CONFIG_ZTEST=y
-
-CONFIG_NETWORKING=y
-CONFIG_NET_L2_ETHERNET=y
-CONFIG_NET_IPV4=y
-CONFIG_NET_IPV6=n
-CONFIG_NET_ARP=y
-CONFIG_NET_UDP=y
-CONFIG_NET_BUF_LOG=y
-CONFIG_NET_LOG=y
-
-CONFIG_ENTROPY_GENERATOR=y
-CONFIG_TEST_RANDOM_GENERATOR=y
