--- tests/net/net_pkt_new/testcase.yaml
@@ -1,7 +1,0 @@
-common:
-  depends_on: netif
-  platform_whitelist: native_posix qemu_x86 qemu_cortex_m3
-tests:
-  net.packet:
-    min_ram: 20
-    tags: net
